[Jones3D] Add auto aim reticle
This adds a new feature missing from the original game; auto aim reticle
The reticle design was inspired by N64 port  of the game and `The Legend of Zelda (N64)` game.
It supports adjustments for color, size, opacity and animation parameters, and integrates
with the auto-aim system for improved gameplay.